+++ projects/docs/src/app/demo-routes.ts
@@ -1,5 +1,4 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
+import { Routes } from '@angular/router';
 import { BasicPageComponent } from './components/basic-example/basic-page.component';
 import { AdvancedPageComponent } from './components/advanced-example/advanced-page.component';
 import { CustomPageComponent } from './components/custom-template-example/custom-page.component';
@@ -7,7 +6,7 @@ import { ServerPageComponent } from './components/server-example/server-page.com
 import { DocumentationPageComponent } from './components/documentation/documentation-page.component';
 
 
-const routes: Routes = [
+export const routes: Routes = [
   {path: '', pathMatch: 'full', component: DocumentationPageComponent},
   {path: 'basic', component: BasicPageComponent},
   {path: 'advanced', component: AdvancedPageComponent},
@@ -15,8 +14,3 @@ const routes: Routes = [
   {path: 'server-paging', component: ServerPageComponent}
 ];
 
-@NgModule({
-    imports: [RouterModule.forRoot(routes)],
-    exports: [RouterModule]
-})
-export class DemoRoutingModule { }