+++ projects/ngx-pagination/src/lib/pagination-controls.component.spec.ts
@@ -16,9 +16,10 @@ describe('PaginationControlsComponent:', () => {
 
     beforeEach(() => {
         TestBed.configureTestingModule({
-            declarations: [PaginationControlsComponent, PaginationControlsDirective, ComponentTestComponent, PaginatePipe],
-            providers: [PaginationService, {provide: LOCALE_ID, useValue: 'en_US' }],
-        });
+    imports: [PaginationControlsComponent, PaginationControlsDirective, PaginatePipe],
+    declarations: [ComponentTestComponent],
+    providers: [PaginationService, { provide: LOCALE_ID, useValue: 'en_US' }],
+});
     });
 
     it('should display the correct page links (simple)', fakeAsync(() => {
@@ -47,9 +48,10 @@ describe('PaginationControlsComponent:', () => {
 
     it('should display the correct page links (formatted numbers over 1000) with dot', fakeAsync(() => {
         TestBed.configureTestingModule({
-            declarations: [PaginationControlsComponent, PaginationControlsDirective, ComponentTestComponent, PaginatePipe],
-            providers: [PaginationService, {provide: LOCALE_ID, useValue: 'de_DE' }],
-        });
+    imports: [PaginationControlsComponent, PaginationControlsDirective, PaginatePipe],
+    declarations: [ComponentTestComponent],
+    providers: [PaginationService, { provide: LOCALE_ID, useValue: 'de_DE' }],
+});
 
         let fixture = TestBed.createComponent(ComponentTestComponent);
         let instance = fixture.componentInstance;